+++ lib/indica_keyboard.dart
@@ -1,5 +1,9 @@
 
 import 'indica_keyboard_platform_interface.dart';
+export 'src/widgets/multilingual_keyboard.dart';
+export 'src/models/keyboard_layout.dart';
+export 'src/services/keyboard_controller.dart';
+export 'src/constants/keyboard_constants.dart';
 
 class IndicaKeyboard {
   Future<String?> getPlatformVersion() {